requirements.txt
web3==6.11.3
requests==2.31.0
pandas==2.1.3
matplotlib==3.8.2
plotly==5.18.0
python-dotenv==1.0.0
eth-account==0.10.0